+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/toolbar/TabCounterToolbarButton.kt
@@ -32,6 +32,7 @@ open class TabCounterToolbarButton(
     private val showTabs: () -> Unit,
     private val store: BrowserStore,
     private val menu: TabCounterMenu? = null,
+    private val showMaskInPrivateMode: Boolean = false,
 ) : Toolbar.Action {
 
     private var reference = WeakReference<TabCounter>(null)
@@ -70,6 +71,8 @@ open class TabCounterToolbarButton(
             )
 
             contentDescription = parent.context.getString(R.string.mozac_feature_tabs_toolbar_tabs_button)
+
+            toggleCounterMask(showMaskInPrivateMode && isPrivate(store))
         }
 
         // Set selectableItemBackgroundBorderless