+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/toolbar/ToolbarIntegration.kt
@@ -138,6 +138,7 @@ class DefaultToolbarIntegration(
             },
             store = store,
             menu = tabCounterMenu,
+            showMaskInPrivateMode = context.settings().feltPrivateBrowsingEnabled,
         )
 
         val tabCount = if (isPrivate) {